use ws adapter properly
